requirements.txt
pytest==7.3.1
Appium-Python-Client==2.10.0
selenium==4.9.0
